app/requirements.txt
pyinstaller==5.3
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
pydicom==2.3.0
PySide6==6.3.1
scikit_learn==1.1.1
scipy==1.9.0
